SC=2763 UX review changes. Moved AddContentButton.vue to organisms. fixed unit tests
