+++ src/components/legacy/TheTopBar.vue
@@ -1,5 +1,5 @@
 <template>
-	<div v-if="showTopbar" class="topbar">
+	<div class="topbar">
 		<!-- ACTIONS -->
 		<div
 			v-if="!fullscreenMode"
@@ -120,11 +120,6 @@ export default {
 	data() {
 		// This solely exists to appear in the coverage report
 		return {};
-	},
-	computed: {
-		showTopbar() {
-			return !window.location.search.includes("inline=1");
-		},
 	},
 	methods: {
 		sendEvent(eventName) {
