+++ src/components/molecules/ContentSearchbar.unit.js
@@ -44,7 +44,7 @@ describe("@components/molecules/ContentSearchbar", () => {
 	});
 
 	it("Switches to active input and clears value when ESCAPE key is pressed", async () => {
-		const wrapper = mount(ContentSearchbar, { attachToDocument: true });
+		const wrapper = mount(ContentSearchbar);
 		const testValue = "testValue";
 		wrapper.setData({ isActive: false, inputValue: testValue });
 		expect(wrapper.vm.isActive).toBe(false);
@@ -57,7 +57,7 @@ describe("@components/molecules/ContentSearchbar", () => {
 	});
 
 	it("Switches to active input and clears value when BACKSPACE key is pressed", async () => {
-		const wrapper = mount(ContentSearchbar, { attachToDocument: true });
+		const wrapper = mount(ContentSearchbar);
 		const testValue = "testValue";
 		wrapper.setData({ isActive: false, inputValue: testValue });
 		expect(wrapper.vm.isActive).toBe(false);
